change functions to shared and moved to @ folder
