+++ imports/ui/components/PagesList.jsx
@@ -11,7 +11,7 @@ import {
   Text,
 } from '@chakra-ui/react';
 
-import { parseTitle } from '../functions';
+import { parseTitle } from '../@/shared';
 import { ScreenClassRender } from 'react-grid-system';
 
 const PagesList = withRouter(({ pageTitles, activePageTitle, history }) => {
