+++ imports/ui/pages/admin/Settings.jsx
@@ -25,7 +25,7 @@ import Loader from '../../components/Loader';
 import Template from '../../components/Template';
 import ListMenu from '../../components/ListMenu';
 import { message, Alert } from '../../components/message';
-import { call, resizeImage, uploadImage } from '../../functions';
+import { call, resizeImage, uploadImage } from '../../@/shared';
 import { adminMenu } from '../../constants/general';
 import SettingsForm from './SettingsForm';
 import FileDropper from '../../components/FileDropper';
