+++ imports/ui/pages/auth/LoginPage.jsx
@@ -5,7 +5,7 @@ import { Box, Center, Heading, Link as CLink, Text } from '@chakra-ui/react';
 import { StateContext } from '../LayoutContainer';
 import Template from '../components/Template';
 import { Login } from './index';
-import { loginWithPassword } from './functions';
+import { loginWithPassword } from './@/shared';
 
 function LoginPage() {
   const { currentUser } = useContext(StateContext);
