+++ imports/ui/pages/pages/Page.js
@@ -10,7 +10,7 @@ import PagesList from '../../components/PagesList';
 import Loader from '../../components/Loader';
 import Template from '../../components/Template';
 
-import { parseTitle } from '../../functions';
+import { parseTitle } from '../../@/shared';
 
 const publicSettings = Meteor.settings.public;
