+++ imports/ui/pages/processes/NewProcess.jsx
@@ -17,11 +17,11 @@ import {
 import { InfoIcon } from '@chakra-ui/icons';
 
 import ProcessForm from '../../components/ProcessForm';
-import { call } from '../../functions';
+import { call } from '../../@/shared';
 import Loader from '../../components/Loader';
 import Template from '../../components/Template';
 import { message, Alert } from '../../components/message';
-import { resizeImage, uploadImage } from '../../functions';
+import { resizeImage, uploadImage } from '../../@/shared';
 import { StateContext } from '../../LayoutContainer';
 
 const successCreation = () => {
