+++ imports/ui/pages/works/NewWork.jsx
@@ -7,7 +7,7 @@ import { StateContext } from '../../LayoutContainer';
 import WorkForm from '../../components/WorkForm';
 import Template from '../../components/Template';
 import { message, Alert } from '../../components/message';
-import { call, resizeImage, uploadImage } from '../../functions';
+import { call, resizeImage, uploadImage } from '../../@/shared';
 
 const formModel = {
   title: '',
